Updates sandbox and controlplane SDK generation
Updates the Makefile to correctly copy the openapi.yml files for sandbox and controlplane SDK generation.
This fixes an issue where the generated SDKs were not up-to-date with the API definitions.

Additionally, marks the generated types file as eslint-disable to avoid linting errors.
The 'path' property in SuccessResponse is now optional.