Adds opt-out mechanism for usage tracking
Implements a mechanism to disable usage tracking based on a setting in the config file or an environment variable.

This allows users to opt-out of data collection. It checks for a `tracking` property in `config.yaml` and falls back to the `BL_TRACKING` environment variable. If neither is set, tracking defaults to enabled.