Implements SDK error tracking with Sentry (#204)
* Implements SDK error tracking with Sentry

Adds Sentry integration for internal SDK error tracking, providing insights into SDK-related issues without interfering with user's own Sentry setup.

- Introduces an isolated Sentry client to capture exceptions originating from the SDK.
- Automatically captures uncaught exceptions and unhandled rejections.
- Adds build information (version, commit, sentry DSN) to identify the SDK version and build.
- Changes the release workflow to differentiate between production, staging, and development releases, with corresponding versioning and publish targets.

* Fix CI

* Adds error logging to Sentry

Ensures errors are captured in Sentry for better monitoring and debugging by explicitly logging errors.

* Filter not wanting environment created by other companies checking packages

* Adds opt-out mechanism for usage tracking

Implements a mechanism to disable usage tracking based on a setting in the config file or an environment variable.

This allows users to opt-out of data collection. It checks for a `tracking` property in `config.yaml` and falls back to the `BL_TRACKING` environment variable. If neither is set, tracking defaults to enabled.